+++ include/IntervalInterrupt_STM32.h
@@ -168,6 +168,6 @@ inline __attribute__((always_inline)) void IntervalInterrupt<T>::stop()
 }
 
 template <Timer T>
-const unsigned long int IntervalInterrupt<T>::FREQ = F_CPU;
+const uint32_t IntervalInterrupt<T>::FREQ = F_CPU;
 
 #endif
